// NodeJS/11_LowLevelDesign/05_DesignProblem/ParkingLotDesignCode/Enum.ts
export enum VehicleType {
  CAR = "CAR",
  MOTORCYCLE = "MOTORCYCLE",
  BICYCLE = "BICYCLE",
  TRUCK = "TRUCK",
  BUS = "BUS"
}

export enum VehiclePass {
  DAILY = "DAILY",
  WEEKLY = "WEEKLY",
  MONTHLY = "MONTHLY",
  NONE = "NONE"
}


export enum ParkingSpotSize {
  SMALL = "SMALL",
  COMPACT = "COMPACT",
  MEDIUM = "MEDIUM",
  BIG = "BIG",
}